+++ max_div/_cli/__init__.py
@@ -4,3 +4,4 @@ from ._cmd_benchmark_internal import *
 from ._cmd_benchmark_platform import *
 from ._cmd_benchmark_solver import *
 from ._cmd_numba_status import *
+from ._cmd_solve import *