+++ CHANGELOG.md
@@ -7,7 +7,9 @@
 
 - **improved**:
     - ci: show slowest tests in github actions when running tests or generating coverage (badge generation) 
-    - cli: add `benchmark platform` command to benchmark platform speed
+    - cli: 
+        - add `benchmark platform` command to benchmark platform speed
+        - add `solve` command for easy test-running the solver on test problems
     - initialization strategies: update internal calling API in preparation of improved progress reporting
 
 <!------------------------------------------------------------------------------------------------->